feat: :sparkles: Math functions
