+++ lib/Makefile
@@ -21,7 +21,7 @@ PKGS     := ft_string dict system ft_math list error
 ft_stringV  := new util util2 array array2 split join
 dictV    := new del util hash set get drop expand print status
 systemV  := alloc read write gnl
-ft_mathV := math compare
+ft_mathV := math compare check
 listV    := list
 errorV   := error
