+++ lib/src/ft_math/compare.c
@@ -19,15 +19,3 @@ int	ft_max(int a, int b)
 		return (a);
 	return (b);
 }
-
-bool	is_digit(char c)
-{
-	return (c >= '0' && c <= '9');
-}
-
-bool	is_overflow(int n, int sign)
-{
-	return (
-		(sign == -1 && n > 0)
-		|| (sign == 1 && n < 0));
-}
